Django/G29/requirement.txt
django
django-cors-headers
djangorestframework
pyjwt
Pillow
